requirements.txt
torch
torchvision
pillow
numpy
tqdm
scikit-optimize
scikit-learn
